+++ packages/terra-form-select/src/_Frame.jsx
@@ -126,6 +126,7 @@ class Frame extends React.Component {
     this.handleKeyDown = this.handleKeyDown.bind(this);
     this.handleMouseDown = this.handleMouseDown.bind(this);
     this.handleInputMouseDown = this.handleInputMouseDown.bind(this);
+    this.visuallyHiddenComponent = React.createRef();
   }
 
   componentDidUpdate(previousProps, previousState) {
@@ -164,6 +165,7 @@ class Frame extends React.Component {
       onMouseDown: this.handleInputMouseDown,
       type: 'text',
       'aria-label': 'search',
+      'aria-disabled': disabled,
       className: cx('search-input', { 'is-hidden': Util.shouldHideSearch(this.props, this.state) }),
     };
 
@@ -395,11 +397,11 @@ class Frame extends React.Component {
     return (
       <div
         {...customProps}
-        role="combobox"
-        aria-controls={this.state.isOpen ? 'terra-select-dropdown' : undefined}
+        role={!disabled ? 'combobox' : undefined}
+        aria-controls={!disabled && this.state.isOpen ? 'terra-select-dropdown' : undefined}
         aria-disabled={!!disabled}
-        aria-expanded={!!this.state.isOpen}
-        aria-haspopup="true"
+        aria-expanded={!!disabled && !!this.state.isOpen}
+        aria-haspopup={!disabled ? 'true' : undefined}
         aria-owns={this.state.isOpen ? 'terra-select-dropdown' : undefined}
         className={selectClasses}
         onBlur={this.handleBlur}
@@ -410,12 +412,13 @@ class Frame extends React.Component {
         ref={(select) => { this.select = select; }}
       >
         {/* eslint-disable-next-line jsx-a11y/interactive-supports-focus */}
-        <div className={cx('display')} role="textbox" onMouseDown={this.openDropdown}>
+        <div role="textbox" aria-disabled={!!disabled} className={cx('display')} onMouseDown={this.openDropdown}>
           {this.getDisplay()}
         </div>
         <div className={cx('toggle')} onMouseDown={this.toggleDropdown}>
           <span className={cx('arrow-icon')} />
         </div>
+        <span className={cx('visually-hidden-component')} ref={this.visuallyHiddenComponent} aria-live="polite" aria-relevant="additions text" aria-atomic="true" />
         {this.state.isOpen
           && (
           <Dropdown
@@ -435,6 +438,7 @@ class Frame extends React.Component {
                  onDeselect,
                  optionFilter,
                  noResultContent,
+                 visuallyHiddenComponent: this.visuallyHiddenComponent,
                  onSelect: this.handleSelect,
                  onRequestClose: this.closeDropdown,
                  searchValue: this.state.searchValue,